third_party/blink/web_tests/fast/events/window-onerror-07-expected.txt

CONSOLE ERROR: Uncaught Error: An exception
This tests that 'window.onerror' is called on the window object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS thisInsideOnerror is window
PASS successfullyParsed is true

TEST COMPLETE
